requirements.txt
google-generativeai==0.5.4
Pillow==10.2.0
folium==0.20.0
Flask==3.0.0 
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai
Pillow
pandas
python-dotenv
gunicorn